cleanup setUp and tearDown
